+++ python/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rmesh"
+name = "rmesh-python"
 version = "0.1.0"
 edition = "2021"
 
@@ -9,13 +9,11 @@ name = "rmesh"
 crate-type = ["cdylib"]
 
 [dependencies]
-ahash = "0.8.11"
 anyhow = "1.0.98"
-itertools = "0.14.0"
-rayon = "1.10.0"
-nalgebra = { version = "0.33.2", features=["bytemuck"] }
+pyo3 = { version = "0.24.0", features=["anyhow"] }
+
 bytemuck = { version = "1.22.0", features = ["derive", "extern_crate_alloc"] }
-cache_access = { path = "./macros/cache_access" }
+rmesh = { path = ".." }
 
 [dev-dependencies]
 approx = "0.5.1"
